Semi-working version. Issues with group deviations.
Current version should handle non-group deviations well, including
cache. There are issues with group member deviations, though -
if a group member has a deviation for a module where the user has
a group submission, then the user gets the same deviations as well,
even if they haven't submitted that said exercise in a group.
This happens even for exercises where the maximum group size is 1.
The problematic queries are in

deviations/models.py:SubmissionRuleDeviationManager.get_max_deviations.

Properly clearing cache is troublesome as well for group submissions.
If a group member's deviation changes, it's not currently reflected on
the user's deadline / max submissions.

Tests also still need to be updated so there are test cases for
module deviations.